requirements.txt
flask==2.3.3
scikit-learn==1.3.0
pillow==10.0.0
numpy==1.26.4
werkzeug==2.3.7
